+++ datatig/exceptions.py
@@ -1,2 +1,6 @@
 class DuplicateRecordIdException(Exception):
     pass
+
+
+class SiteConfigurationException(Exception):
+    pass